Finish the Login rewrite, sort of add support for auth servers Still no authentication support, but the pieces are in place.
Next, the Configuration and Play states, are the hard parts. How to manage sending and receiving players going between worlds, how to transfer components, etc. Thankfully components are already serializable, so the hardest part is already done. Just need to figure out exactly *when* to put people in a world and how to pull them back.